vist-app/requirements.txt
alembic==1.9.0
anyio==3.6.2
asgiref==3.5.2
async-generator==1.10
asyncpg==0.27.0
attrs==22.1.0
certifi==2022.9.24
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cryptography==45.0.6
csv-generator==1.0.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.1
djangorestframework==3.14.0
dnspython==2.7.0
email_validator==2.2.0
envparse==0.2.0
exceptiongroup==1.0.4
fastapi==0.88.0
filelock==3.8.0
greenlet==3.1.1
h11==0.14.0
httpie==3.2.1
idna==3.4
PyJWT==2.8.0
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.3
numpy==1.24.0
outcome==1.2.0
pandas==1.5.2
Pillow==9.2.0
pipenv==2022.12.19
platformdirs==2.5.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.22
pycrypto==2.6.1
pydantic==1.10.4
Pygments==2.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
selenium==4.6.0
simple-crypt==4.1.7
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.45
sqlparse==0.4.2
starlette==0.22.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.20.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
wsproto==1.2.0
